Smooth keyboard scrolling and keyboard navigation
Overhauled keyboard scrolling method.
Instead of using repeated keydown event to scroll by a fixed amount, the new method uses both keydown and keyup to keep track of when a key is being held.

Keyboard navigation is also similarly improved with the smooth scrolling when highlighting elements at the edge of viewport.